Safari-611.1.10/LayoutTests/imported/w3c/web-platform-tests/css/css-display/display-with-float-expected.txt


FAIL computed style for float assert_equals: expected "left" but got "none"
